requirements.txt
pandas>=2.0.0
python-dateutil>=2.8.2
google-cloud-bigquery>=3.11.4
google-auth>=2.20.0
flask
pyarrow>=11.0.0
fastparquet>=2024.1.1
google-cloud-storage>=2.10.0
